// sirius-frontend/src/types/ProjectDayActions.ts
export type ProjectDayAction =
    | "sendAvailabilityCheck"
    | "confirmPerformers"
    | "blacklistPerformers"
    | "createCallSheet"
    | "createFittingSheet"
    | "exportCallSheet"
    | "exportFittingSheet"
    | "exportPerformerList"
    | "massEmail"
    | "sendCallTime"
    | "editProductionDay"
    | "archiveProductionDay"
    | "deleteProductionDay";